fix: :art: split code into small components
